+++ ext/um/um.h
@@ -123,5 +123,8 @@ VALUE um_close(struct um *machine, int fd);
 VALUE um_accept(struct um *machine, int fd);
 VALUE um_accept_each(struct um *machine, int fd);
 VALUE um_socket(struct um *machine, int domain, int type, int protocol, uint flags);
+VALUE um_connect(struct um *machine, int fd, const struct sockaddr *addr, socklen_t addrlen);
+VALUE um_send(struct um *machine, int fd, VALUE buffer, int len, int flags);
+VALUE um_recv(struct um *machine, int fd, VALUE buffer, int maxlen, int flags);
 
 #endif // UM_H